Remove use of cli `distrib_uri` in `publish`
The command line argument `distrib_uri` has two purposes
- for delegates
- to let dune-release in `dune-relesae opam` know where the distribution
tarball is on the web when having published the distribution independently

Furthermore, `dune-release publish` needs to know on which github url
to publish the release. Therefore it inferres the user and repo name from
the dev-repo url. Before this commit, the user and repo name were instead
inferred from the `distrib_uri` when provided, assuming that it is the
standard release uri of a github realease without checking that. This
commit removes this use of `ditrib_uri`.