Add a check command
The new check command allows the user to make sure that the whole release
process will be smooth: it checks if the opam-file is dune-release compatible
and - unless stated otherwise by the
user via the skip options - performs the checks that will be done on the
distribution archive: if it builds, if it passes its tests and if it
passes the dune-release lints. For all that, the same options are taken
into account as will be when releasing.