Lint summary: mention name of the linted package
Only mentioning the directory of the project is misleading, since the lint
doesn't only check if the standard files in the directory exist, but also
lints the opam file of the package. Before this commit, in multi pckage
projects, the lint summary statement for every package was the same - only
the number of lint errors possibly changed.